Updated Cpath_topoSRTM1, cget_latlon, and CompareTopo_AlongPath. HGT file readout done, need work on c++ plotting. This is harder than I thought since there's no simple c++ plotting library, most requires dependencies on openGL and GLUT.

+++ ccode/main/Cpath_topoSRTM1.hpp
@@ -8,6 +8,6 @@ const double pi = 3.14159;
 const double rad2deg = 180/pi;
 const double deg2rad = pi/180;
 
-std::vector<double> returnMin(std::vector<double>, double);
+std::vector<int> returnMin(std::vector<int>, int);
 std::tuple<std::vector<double>, std::vector<double>, std::vector<double>> Cpath_topoSRTM1(std::vector<double>, std::vector<double>);
 #endif
